Combine lab1_list1 and lab1_list2 to one project lab1
