+++ lab1/Test/Test.h
@@ -1,0 +1,16 @@
+#pragma once
+#include <vector>
+#include "../Test/unit_test_framework.h"
+#include "../lab1/HeaderFiles/Containers/Containers.h"
+#include "../lab1/HeaderFiles/Date/Date.h"
+
+void TestStack();
+void TestQueue();
+void TestDeque();
+void TestDateMinusDate();
+void TestDateAddDifference();
+void TestDateMinusDifference();
+void TestWeekDay();
+void TestWeekInMonth();
+void TestWeekInYear();
+void TestTimeManagmentInContainers();
