+++ lab1/lab1.sln
@@ -1,0 +1,41 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29215.179
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "lab1", "lab1\lab1.vcxproj", "{4348EECF-658F-4725-9536-C2DC439413FD}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Test", "Test\Test.vcxproj", "{C496B637-EF1F-41EF-B99E-A476F6C11788}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{4348EECF-658F-4725-9536-C2DC439413FD}.Debug|x64.ActiveCfg = Debug|x64
+		{4348EECF-658F-4725-9536-C2DC439413FD}.Debug|x64.Build.0 = Debug|x64
+		{4348EECF-658F-4725-9536-C2DC439413FD}.Debug|x86.ActiveCfg = Debug|Win32
+		{4348EECF-658F-4725-9536-C2DC439413FD}.Debug|x86.Build.0 = Debug|Win32
+		{4348EECF-658F-4725-9536-C2DC439413FD}.Release|x64.ActiveCfg = Release|x64
+		{4348EECF-658F-4725-9536-C2DC439413FD}.Release|x64.Build.0 = Release|x64
+		{4348EECF-658F-4725-9536-C2DC439413FD}.Release|x86.ActiveCfg = Release|Win32
+		{4348EECF-658F-4725-9536-C2DC439413FD}.Release|x86.Build.0 = Release|Win32
+		{C496B637-EF1F-41EF-B99E-A476F6C11788}.Debug|x64.ActiveCfg = Debug|x64
+		{C496B637-EF1F-41EF-B99E-A476F6C11788}.Debug|x64.Build.0 = Debug|x64
+		{C496B637-EF1F-41EF-B99E-A476F6C11788}.Debug|x86.ActiveCfg = Debug|Win32
+		{C496B637-EF1F-41EF-B99E-A476F6C11788}.Debug|x86.Build.0 = Debug|Win32
+		{C496B637-EF1F-41EF-B99E-A476F6C11788}.Release|x64.ActiveCfg = Release|x64
+		{C496B637-EF1F-41EF-B99E-A476F6C11788}.Release|x64.Build.0 = Release|x64
+		{C496B637-EF1F-41EF-B99E-A476F6C11788}.Release|x86.ActiveCfg = Release|Win32
+		{C496B637-EF1F-41EF-B99E-A476F6C11788}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {EBF36AEC-BF7E-4564-BBB0-D876DD955096}
+	EndGlobalSection
+EndGlobal
