+++ lab1/lab1/HeaderFiles/Containers/Containers.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "Deque.h"
+#include "Queue.h"
+#include "Stack.h"
