+++ lab1/lab1/lab1.vcxproj
@@ -20,8 +20,8 @@
   </ItemGroup>
   <PropertyGroup Label="Globals">
     <VCProjectVersion>16.0</VCProjectVersion>
-    <ProjectGuid>{949EBAE9-ACCA-4D7E-A220-E2AFEE890BD1}</ProjectGuid>
-    <RootNamespace>lab1list1v3</RootNamespace>
+    <ProjectGuid>{4348EECF-658F-4725-9536-C2DC439413FD}</ProjectGuid>
+    <RootNamespace>lab1</RootNamespace>
     <WindowsTargetPlatformVersion>10.0</WindowsTargetPlatformVersion>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
@@ -123,15 +123,19 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="demontration.cpp" />
-    <ClCompile Include="Source.cpp" />
-    <ClCompile Include="test.cpp" />
+    <ClCompile Include="CppFiles\Demonstration.cpp" />
+    <ClCompile Include="CppFiles\Overload.cpp" />
+    <ClCompile Include="CppFiles\Time_is_now.cpp" />
+    <ClCompile Include="CppFiles\Time_managment.cpp" />
+    <ClCompile Include="main.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="containers.h" />
-    <ClInclude Include="demonstration.h" />
-    <ClInclude Include="test.h" />
-    <ClInclude Include="unit_test_framework.h" />
+    <ClInclude Include="HeaderFiles\Containers\Containers.h" />
+    <ClInclude Include="HeaderFiles\Containers\Deque.h" />
+    <ClInclude Include="HeaderFiles\Containers\Queue.h" />
+    <ClInclude Include="HeaderFiles\Containers\Stack.h" />
+    <ClInclude Include="HeaderFiles\Date\Date.h" />
+    <ClInclude Include="HeaderFiles\Demonstration\Demonstration.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
