--- lab1/lab1_list1_v3/lab1_list1_v3.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29215.179
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "lab1_list1_v3", "lab1_list1_v3\lab1_list1_v3.vcxproj", "{949EBAE9-ACCA-4D7E-A220-E2AFEE890BD1}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{949EBAE9-ACCA-4D7E-A220-E2AFEE890BD1}.Debug|x64.ActiveCfg = Debug|x64
-		{949EBAE9-ACCA-4D7E-A220-E2AFEE890BD1}.Debug|x64.Build.0 = Debug|x64
-		{949EBAE9-ACCA-4D7E-A220-E2AFEE890BD1}.Debug|x86.ActiveCfg = Debug|Win32
-		{949EBAE9-ACCA-4D7E-A220-E2AFEE890BD1}.Debug|x86.Build.0 = Debug|Win32
-		{949EBAE9-ACCA-4D7E-A220-E2AFEE890BD1}.Release|x64.ActiveCfg = Release|x64
-		{949EBAE9-ACCA-4D7E-A220-E2AFEE890BD1}.Release|x64.Build.0 = Release|x64
-		{949EBAE9-ACCA-4D7E-A220-E2AFEE890BD1}.Release|x86.ActiveCfg = Release|Win32
-		{949EBAE9-ACCA-4D7E-A220-E2AFEE890BD1}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {CBA1828C-7AA5-4A71-BEBF-C7EC557174AF}
-	EndGlobalSection
-EndGlobal
