--- lab1/lab1_list1_v3/lab1_list1_v3/test.h
@@ -1,7 +1,0 @@
-#pragma once
-#include "unit_test_framework.h"
-#include "containers.h"
-
-void TestStack();
-void TestQueue();
-void TestDeque();
