--- lab1/lab1_list2_v9/lab1_list2_v9.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29215.179
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "lab1_list2_v9", "lab1_list2_v9\lab1_list2_v9.vcxproj", "{1B2121FE-57C8-410C-B975-BBF60D22AE5D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1B2121FE-57C8-410C-B975-BBF60D22AE5D}.Debug|x64.ActiveCfg = Debug|x64
-		{1B2121FE-57C8-410C-B975-BBF60D22AE5D}.Debug|x64.Build.0 = Debug|x64
-		{1B2121FE-57C8-410C-B975-BBF60D22AE5D}.Debug|x86.ActiveCfg = Debug|Win32
-		{1B2121FE-57C8-410C-B975-BBF60D22AE5D}.Debug|x86.Build.0 = Debug|Win32
-		{1B2121FE-57C8-410C-B975-BBF60D22AE5D}.Release|x64.ActiveCfg = Release|x64
-		{1B2121FE-57C8-410C-B975-BBF60D22AE5D}.Release|x64.Build.0 = Release|x64
-		{1B2121FE-57C8-410C-B975-BBF60D22AE5D}.Release|x86.ActiveCfg = Release|Win32
-		{1B2121FE-57C8-410C-B975-BBF60D22AE5D}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {75A7DB4C-7689-4E85-BA04-3D4A1C26FAE7}
-	EndGlobalSection
-EndGlobal
