--- lab1/lab1_list2_v9/lab1_list2_v9/demonstration.h
@@ -1,13 +1,0 @@
-#pragma once
-#include "containers.h"
-#include "date.h"
-#include "windows.h"
-#include <cstdlib>
-#include <vector>
-
-using namespace std;
-
-int menu();
-void demonstration_mode_stack();
-void demonstration_mode_queue();
-void demonstration_mode_deque();
