--- lab1/lab1_list2_v9/lab1_list2_v9/lab1_list2_v9.vcxproj.filters
@@ -1,60 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <ItemGroup>
-    <Filter Include="Исходные файлы">
-      <UniqueIdentifier>{4FC737F1-C7A5-4376-A066-2A32D752A2FF}</UniqueIdentifier>
-      <Extensions>cpp;c;cc;cxx;def;odl;idl;hpj;bat;asm;asmx</Extensions>
-    </Filter>
-    <Filter Include="Файлы заголовков">
-      <UniqueIdentifier>{93995380-89BD-4b04-88EB-625FBE52EBFB}</UniqueIdentifier>
-      <Extensions>h;hh;hpp;hxx;hm;inl;inc;ipp;xsd</Extensions>
-    </Filter>
-    <Filter Include="Файлы ресурсов">
-      <UniqueIdentifier>{67DA6AB6-F800-4c08-8B7A-83BB121AAD01}</UniqueIdentifier>
-      <Extensions>rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tiff;tif;png;wav;mfcribbon-ms</Extensions>
-    </Filter>
-    <Filter Include="Файлы заголовков\date">
-      <UniqueIdentifier>{c0aecc52-44ee-46d5-8da7-3aa7996ecd52}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Файлы заголовков\containers">
-      <UniqueIdentifier>{fabcfb85-0789-42e7-a668-77f0d9106ce2}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Файлы заголовков\demonstration">
-      <UniqueIdentifier>{e55356cf-83a4-4f20-8cd8-f377fd45bf8a}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Файлы заголовков\test">
-      <UniqueIdentifier>{11d81994-4945-48fa-a025-149c6a9ec282}</UniqueIdentifier>
-    </Filter>
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="main.cpp">
-      <Filter>Исходные файлы</Filter>
-    </ClCompile>
-    <ClCompile Include="date.cpp">
-      <Filter>Файлы заголовков\date</Filter>
-    </ClCompile>
-    <ClCompile Include="demonstration.cpp">
-      <Filter>Файлы заголовков\demonstration</Filter>
-    </ClCompile>
-    <ClCompile Include="test.cpp">
-      <Filter>Файлы заголовков\test</Filter>
-    </ClCompile>
-  </ItemGroup>
-  <ItemGroup>
-    <ClInclude Include="date.h">
-      <Filter>Файлы заголовков\date</Filter>
-    </ClInclude>
-    <ClInclude Include="containers.h">
-      <Filter>Файлы заголовков\containers</Filter>
-    </ClInclude>
-    <ClInclude Include="demonstration.h">
-      <Filter>Файлы заголовков\demonstration</Filter>
-    </ClInclude>
-    <ClInclude Include="test.h">
-      <Filter>Файлы заголовков\test</Filter>
-    </ClInclude>
-    <ClInclude Include="unit_test_framework.h">
-      <Filter>Файлы заголовков\test</Filter>
-    </ClInclude>
-  </ItemGroup>
-</Project>
