--- lab1/lab1_list2_v9/lab1_list2_v9/test.h
@@ -1,12 +1,0 @@
-#pragma once
-#include "unit_test_framework.h"
-#include "containers.h"
-#include "date.h"
-
-void TestDateMinusDate();
-void TestDateAddDifference();
-void TestDateMinusDifference();
-void TestWeekDay();
-void TestWeekInMonth();
-void TestWeekInYear();
-void TestTimeManagmentInContainers();
